+++ src/opensmt/options/SMTConfig.h
@@ -343,7 +343,7 @@ namespace opensmt {
 
     bool usedForInitialization = false; // Some options can be changed only before this config is used for initialization of MainSolver
     bool isPreInitializationOption(const char* o_name) {
-        return strcmp(o_name, o_produce_inter) == 0 || strcmp(o_name, o_produce_proofs) == 0
+        return strcmp(o_name, o_produce_inter) == 0 || strcmp(o_name, o_produce_proofs) == 0 || strcmp(o_name, o_produce_unsat_cores) == 0
           || strcmp(o_name, o_sat_pure_lookahead) == 0 || strcmp(o_name, o_sat_lookahead_split) == 0
           || strcmp(o_name, o_sat_picky) == 0 || strcmp(o_name, o_sat_scatter_split) == 0
           || strcmp(o_name, o_ghost_vars) == 0;